Control.Kernmantle.Error

Description

Deal with throw and try effects

Documentation

class ThrowEffect ex eff where

A class for binary effects that can possibly throw exceptions

Methods

throwE :: eff (Either ex b) b

class TryEffect ex eff where

A class for binary effects that can catch exceptions

Methods

tryE :: eff a b -> eff a (Either ex b)

module Control.Exception.Safe