Control.Kernmantle.Rope

Description

A Rope connects together various effect Strands that get interlaced together.

A Strand is an effet with a parameter and an output. No constraint is placed on the Strand, but once combined in a Rope, that Rope will be an Arrow and a Profunctor. Strands in a Rope are named via labels to limit ambiguity.

An action that targets some Strand can be lifted to a Rope that contains that Strand with the strand function.

Re-exports

module Data.Profunctor.Trans

newtype Cayley (f :: k -> Type) (p :: k1 -> k2 -> k) (a :: k1) (b :: k2)

Static arrows. Lifted by Applicative.

Cayley has a polymorphic kind since 5.6.

Constructors

Cayley

Fields

runCayley :: f (p a b)

Rope

data Rope (record :: RopeRec) (mantle :: [Strand]) (core :: BinEff) a b

Rope is a free arrow built out of _several_ binary effects (ie. effects with kind * -> * -> *). These effects are called Strands, they compose the mantle, can be interpreted in an interp effect and can be interlaced "on top" of an existing core effect.

type TightRope = Rope ARec

A Rope that is "tight", meaning you cannot weave new Strands to it. The strand function is O(1) on TightRopes whatever the number of Strands.

type LooseRope = Rope Rec

A Rope that is "loose", meaning you can weave new Strands to it. The strand function is O(n) on LooseRopes, n being the number of Strands.

tighten :: (RecApplicative m, RPureConstrained (IndexableField m) m) => LooseRope m core :-> TightRope m core

Turn a LooseRope into a TightRope

loosen :: NatToInt (RLength m) => TightRope m core :-> LooseRope m core

Turn a TightRope into a LooseRope. This is very often the first step in a chain of weaves.

mapRopeCore :: (core a b -> core a' b') -> Rope r m core a b -> Rope r m core a' b'

Applies a function on the core action of the Rope

mergeStrands :: Label n1 -> Label n2 -> LooseRope ('(n1, binEff) ': ('(n2, binEff) ': mantle)) core :-> LooseRope ('(n1, binEff) ': mantle) core

Merge two strands that have the same effect type. Keeps the first name.

Binary Effects and Strands

type BinEff = * -> * -> *

The kind for all binary effects. First param is usually an input (contravariant) of the effect and second one an output (covariant).

type Strand = (Symbol, BinEff)

The kind for a named binary effect. Must remain a tuple because that's what vinyl expects.

type (:::) (a :: k) (b :: k1) = '(a, b)

Alias for Field spec

type family StrandName t where ...

Equations

StrandName '(name, eff) = name

type family StrandEff t where ...

Equations

StrandEff '(name, eff) = eff

type family InRope l eff rope where ...

Tells that a strand of effects '(l,eff) is present in the rope

Equations

InRope l eff (Rope record mantle core) = (HasField record l mantle mantle eff eff, RecElemFCtx record (Weaver core))

type AllInMantle strands mantle core = (NatToInt (RLength mantle), RecApplicative mantle, RPureConstrained (IndexableField mantle) mantle, LooseRope mantle core `Entwines` strands, TightRope mantle core `Entwines` strands)

Useful for functions that want to use tighten/loosen

strand :: InRope l eff (Rope r m c) => Label l -> eff :-> Rope r m c

mapStrand :: InRope l eff (Rope r m c) => Label l -> (eff :-> eff) -> Rope r m c :-> Rope r m c

Rope Interpretation

weave

Arguments

:: forall name binEff mantle core. Label name
The name of the strand to be woven.

-> ((LooseRope ('(name, binEff) ': mantle) core :-> core) -> binEff :-> core)
The interpretation function for the effect, which can depend on the global interpretation function for the rope.

-> LooseRope ('(name, binEff) ': mantle) core :-> LooseRope mantle core
An interpretation function for a Rope with containing the given effect strand, transformed into the same Rope but with that effect woven in the core.

Adds a new effect strand in the Rope by decribing how to interpret that in the core. The interpretation function is fed the fixpoint of all interpretation functions, meaning the interpretation function can itself use effects in the Rope. For example, given an effect Eff, whose label is "effect" being interpreted in a core effect CoreEff, one first defines an interpretation function:

  interpret ::
    (LooseRope '("effect",Eff) ': mantle) core :-> CoreEff) ->
     Eff a b -> CoreEff a b

  interpret ::
    (LooseRope '("effect",Eff) ': mantle) core :-> CoreEff) ->
    (Eff :-> CoreEff)

where p :-> q corresponds to forall a b. p a b -> q a b, and obtain an interpreter for the strand:

  strandInterpreter ::
    LooseRope ('("effect",Eff) ': mantle) CoreEff :->
    LooseRope mantle CoreEff
  strandInterpreter = weave #effect interpret

weave'

Arguments

:: forall name binEff mantle core. Label name
The name of the strand to be woven.

-> (binEff :-> core)
The interpretation function for the effect.

-> LooseRope ('(name, binEff) ': mantle) core :-> LooseRope mantle core
An interpretation function for a Rope with containing the given effect strand, transformed into the same Rope but with that effect woven in the core.

A version of weave where the strand can be directly interpreted in the core with no need to trigger other effects. For example, given an effect Eff, whose label is "effect" being interpreted in a core effect CoreEff, one first defines an interpretation function:

  interpret :: Eff a b -> CoreEff a b
  interpret :: Eff :-> CoreEff

and obtain an interpreter for the strand:

  strandInterpreter ::
    LooseRope ('("effect",Eff) ': mantle) CoreEff :->
    LooseRope mantle CoreEff
  strandInterpreter = weave' #effect interpret

weaveK :: HasKleisli m core => Label name -> (forall a b. binEff a b -> a -> m b) -> LooseRope ('(name, binEff) ': mantle) core :-> LooseRope mantle core

A shortcut for weave' when your core HasKleisli

class WeaveAll mantle core where

Supports weaveAll over any Rope

Methods

weaveAll :: (forall lbl eff. lbl -> eff :-> core) -> LooseRope mantle core :-> core

Weaves all the remaining strands in the core, provided a polymorphic-enough interpretation function can be provided.

This function is notably useful to implement the "Rope-in-Rope" pattern, when a Rope (possibly wrapped in Cayley) is used as the code for another Rope. weaveAll in that case permits to transfer all the effects of the outer Rope to the core Rope.

retwine :: RetwinableAs r strands core strands' => Rope r strands core :-> Rope r strands' core

Reorders the strands to match some external context. strands' can contain more elements than strands. Note it works on both TightRopes and LooseRopes

untwine :: LooseRope '[] core :-> core

Strips a Rope of its empty mantle. Usually the last step of the interpretation of a Rope.

Predicates

type AnyRopeWith strands coreConstraints a b = forall mantle r core. (Rope r mantle core `Entwines` strands, core `SatisfiesAll` coreConstraints) => Rope r mantle core a b

A AnyRopeWith strands costraints is a Rope with contains the strands in the variable strands , and whose core effect obeys the constraints in the contraints variable.

type TightRopeWith strands coreConstraints a b = forall mantle core. (TightRope mantle core `Entwines` strands, core `SatisfiesAll` coreConstraints) => TightRope mantle core a b

type LooseRopeWith strands coreConstraints a b = forall mantle core. (LooseRope mantle core `Entwines` strands, core `SatisfiesAll` coreConstraints) => LooseRope mantle core a b

type family Entwines rope (strands :: [Strand]) :: Constraint where ...

Tells whether a collection of strands is in a Rope.

Equations

Entwines rope '[] = ()

Entwines rope ('(name, eff) ': strands) = (InRope name eff rope, rope `Entwines` strands)

type family SatisfiesAll (x :: k) (csts :: [k -> Constraint]) :: Constraint where ...

Tells whether a type satisfies all the given constraints.

Equations

SatisfiesAll x '[] = ()

SatisfiesAll x (c1 ': cnsts) = (c1 x, x `SatisfiesAll` cnsts)

Reexports

type (:->) (p :: k -> k1 -> Type) (q :: k -> k1 -> Type) = forall (a :: k) (b :: k1). p a b -> q a b infixr 0

(:->) has a polymorphic kind since 5.6.

data Label (a :: Symbol)

Proxy for label type

fromLabel :: IsLabel x a => a

(&) :: a -> (a -> b) -> b infixl 1

& is a reverse application operator. This provides notational convenience. Its precedence is one higher than that of the forward application operator $, which allows & to be nested in $.

>>> 5 & (+1) & show
"6"

Since: base-4.8.0.0

mkRope :: (record (Weaver core) mantle -> core a b) -> Rope record mantle core a b

runRope :: Rope record mantle core a b -> record (Weaver core) mantle -> core a b