Data.Profunctor.Trans

Description

This module exposes the SieveTrans class and some Sieve transformers based on usual Reader and Writer

Documentation

class SieveTrans f cat | cat -> f where

A general version of Sieve that allows mapping and recursively reaching the sieve

Methods

liftSieve :: (a -> f b) -> cat a b

mapSieve :: ((a -> f b) -> a' -> f b') -> cat a b -> cat a' b'

type HasKleisli = SieveTrans

Just an alias

liftKleisli :: HasKleisli m eff => (a -> m b) -> eff a b

Just an alias

mapKleisli :: HasKleisli m eff => ((a -> m b) -> a' -> m b') -> eff a b -> eff a' b'

Just an alias

type HasKleisliIO m eff = (HasKleisli m eff, MonadIO m)

liftKleisliIO :: HasKleisliIO m eff => (a -> IO b) -> eff a b

When you want to lift some IO action in a Sieve of any MonadIO

type (~>) = Cayley infixr 1

An alias to make signatures more readable

type Reader r = Reader r

type Writer w = Writer w

fmapping :: Functor f => f t -> (t -> eff a b) -> (f ~> eff) a b

mapCayleyEff :: Functor f => (eff a b -> eff' a' b') -> (f ~> eff) a b -> (f ~> eff') a' b'

mapCayley in profunctors maps the functor. mapCayleyEff maps the effect in it.

reading :: (t -> eff a b) -> (Reader t ~> eff) a b

mapReader :: (t' -> eff a b -> (t, eff' a' b')) -> (Reader t ~> eff) a b -> (Reader t' ~> eff') a' b'

mapReader_ :: (t -> eff a b -> eff' a' b') -> (Reader t ~> eff) a b -> (Reader t ~> eff') a' b'

runReader :: t -> (Reader t ~> eff) a b -> eff a b

writing :: w -> eff :-> (Writer w ~> eff)

mapWriter :: (w -> eff a b -> (w', eff' a' b')) -> (Writer w ~> eff) a b -> (Writer w' ~> eff') a' b'

mapWriter_ :: (w -> eff a b -> eff' a' b') -> (Writer w ~> eff) a b -> (Writer w ~> eff') a' b'

runWriter :: (Writer w ~> eff) a b -> (w, eff a b)

runWriter_ :: (Writer w ~> eff) a b -> eff a b

swap :: (a, b) -> (b, a)

returning :: Arrow eff => b -> eff a b

perform :: a -> Kleisli m a b -> m b

Just a flipped variant of runKleisli