Control.Kernmantle.Arrow

Description

Helper functions for Arrow effects

Documentation

parMapA :: ArrowChoice a => a b c -> a [b] [c]

Map an arrow over a list using (***)

seqMapA :: ArrowChoice a => a b c -> a [b] [c]

Map an arrow over a list using (>>>) and first

foldlA :: ArrowChoice a => a (b, acc) acc -> a ([b], acc) acc

Repeats an arrow step in order to fold a list

filterA :: ArrowChoice a => a b Bool -> a [b] [b]

Filter a list given an arrow filter

catchE

Arguments

:: (TryEffect ex eff, ArrowChoice eff)

=> eff e c
The effect to wrap

-> eff (e, ex) c
What to do in case of exception

-> eff e c

If a TryEffect is also an ArrowChoice, then we can implement catch

newtype SplitId

A SplitId is a ratio of two positive numbers

Constructors

SplitId (Ratio Word)

split :: (s -> SplitId) -> (s -> SplitId -> s) -> (s -> a) -> (a -> b -> c) -> (s -> b) -> s -> c

split defines the Calkin-Wilf tree. Its guaranteed never to produce twice the same result if we split repeatedly starting from 1.

data ArrowIdent

The identifiers reflect the structure of the pipeline that led to some task. They are split at every use of (.), (***), (|||) or (+). This makes it so every task in the pipeline has a different identifier.

Constructors

ArrowIdent

Fields

aidChoice :: !SplitId

aidPlus :: !SplitId

aidPar :: !SplitId

aidComp :: !SplitId

newtype AutoIdent arr a b

An arrow transformer that can automatically determine an identifier from its position in a pipeline. It is isomorphic to a Reader ArrowIdent ~> arr, but we need a different Arrow instance than what Cayley provides.

Constructors

AutoIdent (ArrowIdent -> arr a b)

runAutoIdent' :: SplitId -> AutoIdent arr a b -> arr a b

runAutoIdent :: AutoIdent arr a b -> arr a b

class HasAutoIdent wrappedEff eff | eff -> wrappedEff where

All effects that internally feature some AutoIdent

Methods

liftAutoIdent :: (ArrowIdent -> wrappedEff a b) -> eff a b