Control.Kernmantle.Rope.Internal

Documentation

type BinEff = * -> * -> *

The kind for all binary effects. First param is usually an input (contravariant) of the effect and second one an output (covariant).

type UnaryEff = * -> *

The kind for unary effects

type Strand = (Symbol, BinEff)

The kind for a named binary effect. Must remain a tuple because that's what vinyl expects.

type family StrandName t where ...

Equations

StrandName '(name, eff) = name

type family StrandEff t where ...

Equations

StrandEff '(name, eff) = eff

type RopeRec = (Strand -> *) -> [Strand] -> *

The kind for records that will contain Weavers. First type param will most often be Weaver someCore

newtype Weaver (interp :: BinEff) (strand :: Strand)

Runs one strand (* -> * -> * effect) in a interp effect. Is parameterized over a Strand (and not just a BinEffect) even if it ignores its name internally because that's what is expect by the RopeRec

Constructors

Weaver

Fields

weaveStrand :: StrandEff strand :-> interp

mapWeaverInterp :: (interp :-> interp') -> Weaver interp strand -> Weaver interp' strand

newtype RopeRunner (record :: RopeRec) (mantle :: [Strand]) (interp :: BinEff) (core :: BinEff) a b

The internal implementation of a Rope, where interp and core can be two different types (this facilitates implementation of some functions).

record holds the functions to interpret the strands in an interp effect. And then all these interpreted effects will run in a core effect.

Constructors

RopeRunner (record (Weaver interp) mantle -> core a b)

splitRopeRunner :: RopeRunner Rec (mantle1 ++ mantle2) interp core :-> RopeRunner Rec mantle1 interp (RopeRunner Rec mantle2 interp core)

Splits a RopeRunner in two parts, so we can select several strands to act on

type RetwinableAs record strands core strands' = (RecSubset record strands strands' (RImage strands strands'), RecSubsetFCtx record (Weaver core))

Indicates that strands can be reordered and considered a subset of strands'

joinRopeRunner :: (RetwinableAs r mantle1 interp mantle, RetwinableAs r mantle2 interp mantle) => RopeRunner r mantle1 interp (RopeRunner r mantle2 interp core) :-> RopeRunner r mantle interp core

type family MapStrandEffs f mantle where ...

Wrap each strand effect inside a type constructor

Equations

MapStrandEffs f '[] = '[]

MapStrandEffs f (s ': strands) = '(StrandName s, f (StrandEff s)) ': MapStrandEffs f strands

effmapRopeRec :: EffFunctor f => Rec (Weaver interp) strands -> Rec (Weaver (f interp)) (MapStrandEffs f strands)

unwrapRopeRunner :: EffFunctor f => RopeRunner Rec (MapStrandEffs f strands) (f core) core :-> RopeRunner Rec strands core core

When all the strands of a Rope are built by the same EffFunctor, we can run them. See splitRope to isolate some strands in a Rope

unwrapSomeStrands

Arguments

:: (EffFunctor f, RMap (MapStrandEffs f mantle1))

=> (f core' :-> interp)
How to run the extracted EffFunctor layer

-> (RopeRunner Rec mantle2 interp core :-> core')
What to do with the remaining strands (those not wrapped)

-> RopeRunner Rec (MapStrandEffs f mantle1 ++ mantle2) interp core :-> RopeRunner Rec mantle1 core' core'
The resulting RopeRunner, where